Update verify.py
Coordinate values must have a descending order for latitude, and ascending order for longitude. A datacube could have different coordinate values order and still be a correct input or datacube. However, if that is the case, the xcube viewer will not plot the data on the map (even when the time graph works properly). There are two ways of solving this, adapting the viewer to be able of plotting the data whatever is the coordinate values order, or add this few line to xcube verify.