New option to xcube resample
This is a new option to select when using xcube resample. This is related to frequency options. Downsampling along time is very interesting to compare between days, months or years. But, it is missing an option where we can create an overview cube where we have mean, max, min, etc, of the whole dataset. This overview cube will have time dimension equal to 1. This could be achieve without changing anything and using xcube resample with a frequency higher than the period of the whole dataset. However, I think it is nicer if we can add the option since it is just a few lines more. Here, I've changed the resample.py in core folder. But, resamply.py in cli folder must also be changed. Just to inform about this new option. I'll do another pull request for it.