Fix dockerfile such that serving from docker container works
When serving xcubes from inside a docker container the address has to be
set to 0.0.0.0 for it to work. This is not documented anywhere and the
error message was not quite clear. I've added the docker compose file so
all anyone has to do is 'docker-compose up -d' this will spin up the
image in such a way that the xcube-viewer immediately picks up on it.