Remove diff size limit in DumpTest.test_dump_ds
The abovementioned test failed nondeterministically on Travis (i.e.
failed on a commit for which it had previously passed). The cause was
difficult to diagnose because the diff output had been suppressed due
to excessive length. This commit removes the diff size limit for this
test so that its failure can be diagnosed if it occurs again.