Make some Dataset.sel check TZ compatibility
Addresses #605.

This commit adds checks to several Dataset.sel calls in the xcube
codebase to make sure that the timezone-awareness of the selector
matches that of the time variable in the dataset.

Currently a lot of unit tests fail with these changes, so some
further improvements / bugfixes will be needed.