timeindex ndarray indexers: only handle singletons
The previous generalized implementation turned out to be unreliable,
as did a couple of alternative implementations. Instead I implemented
explicit and robust handling of indexers which are 0-dimensional
ndarrays. >0-dimensional indexers are now left unchanged and a
warning is logged; it's not clear to me whether these every occur in
the wild, so it probably makes sense to defer implementation until we
actually encounter one (if ever).