OGC API: change gdc_version to 1.0.0-beta
Previously the root OGC endpoint returned a gdc_version of 1.0.0-beta.1
(to allow for differentiation between different beta implementations)
but "1.0.0-beta" is in fact required by the draft specifications, and
clients may reject other version specifiers.

The conformance URL already lacked the ".1" suffix, so is not changed
by this commit.