Make new_cube set time encoding dtype explicitly
Recent xarray versions raise an error when trying to serialize
a chunked datetime array for which the encoding specifies the units
but not the (encoded) dtype. Cubes generated by new_cube have this
characteristics, causing errors when one attempts to chunk and save
them, and leading to many unit test failures.

This commit adds a time_encoding_dtype parameter to new_cube with
a default value of int64 (suitable for serializing the default
time_dtype of datetime64), and adds it as the value of dtype in
the encoding dictionary of the time and time_bnds variables.

Partially addresses #939.