Web API: add "slash" parameter for routes
Add an optional "slash" parameter to the server.api.Api.route
decorator, a corresponding constructor parameter and field to
ApiRoute, and code to handle this field to
TornadoFramework.add_routes.

Passing slash=True causes the framework to add an additional
second route, handling the same path as the specified route
but with a suffixed "/" character.

Addresses issue #965.