Update MaskSet.__str__ for NumPy 2.0.0
Update the MaskSet.__str__ implementation to explicitly unpack
the mask/value tuple before interpolating it into the string.
Recursive __str__-ification of tuples actually calls __repr__
on the elements (see PEP 3140), but we want __str__. This didn't
matter with NumPy 1 because both __str__ and __repr__ were the same
for the relevant types, but in NumPy 2.0 __repr__ has also includes
the dtype.